requirements.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
scipy
rasterio
Pillow
segmentation_models_pytorch